fix(#46): move shared type modules from .pi/extensions/ to .pi/lib/
Pi auto-loads every .ts file in .pi/extensions/ as an extension factory.
codebase-types, github-types, lsp-types, and types are pure type/utility
modules imported by actual extensions — they don't export a factory function.
Moving them to .pi/lib/ fixes the 'does not export a valid factory function'
startup errors.
