+++ .pi/extensions/codebase-memory.ts
@@ -3,7 +3,7 @@ import { BorderedLoader } from "@earendil-works/pi-coding-agent";
 import { Type } from "typebox";
 import { createHash } from "node:crypto";
 import { homedir } from "node:os";
-import { HEAVY_TOOLS, QUERY_TIMEOUT, HEAVY_TIMEOUT } from "./codebase-types.js";
+import { HEAVY_TOOLS, QUERY_TIMEOUT, HEAVY_TIMEOUT } from "../lib/codebase-types.js";
 
 const BINARY = `${homedir()}/.local/bin/codebase-memory-mcp`;
